@@ -5,10 +5,11 @@
   "license": "MIT",
   "repository": "JustinBeckwith/linkinator",
   "author": "Justin Beckwith",
-  "main": "build/src/index.js",
-  "types": "build/src/index.d.ts",
+  "exports": "./build/src/index.js",
+  "types": "./build/src/index.d.ts",
+  "type": "module",
   "bin": {
-    "linkinator": "build/src/cli.js"
+    "linkinator": "./build/src/cli.js"
   },
   "scripts": {
     "pretest": "npm run compile",
@@ -22,45 +23,45 @@
     "docs-test": "node build/src/cli.js ./README.md"
   },
   "dependencies": {
-    "chalk": "^4.0.0",
+    "chalk": "^4.1.2",
     "escape-html": "^1.0.3",
-    "gaxios": "^4.0.0",
-    "glob": "^7.1.6",
-    "htmlparser2": "^7.1.2",
-    "jsonexport": "^3.0.0",
-    "marked": "^2.0.0",
-    "meow": "^9.0.0",
+    "gaxios": "^4.3.2",
+    "glob": "^7.2.0",
+    "htmlparser2": "^7.2.0",
+    "jsonexport": "^3.2.0",
+    "marked": "^4.0.3",
+    "meow": "^10.1.1",
     "mime": "^3.0.0",
     "server-destroy": "^1.0.1",
-    "update-notifier": "^5.0.0"
+    "update-notifier": "^5.1.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.2.7",
-    "@types/escape-html": "^1.0.0",
-    "@types/glob": "^7.1.3",
-    "@types/marked": "^3.0.0",
-    "@types/meow": "^5.0.0",
+    "@types/chai": "^4.2.22",
+    "@types/escape-html": "^1.0.1",
+    "@types/glob": "^7.2.0",
+    "@types/jsonexport": "^3.0.2",
+    "@types/marked": "^4.0.0",
     "@types/mime": "^2.0.3",
     "@types/mocha": "^9.0.0",
-    "@types/node": "^16.0.0",
-    "@types/server-destroy": "^1.0.0",
-    "@types/sinon": "^10.0.0",
-    "@types/update-notifier": "^5.0.0",
-    "c8": "^7.0.0",
-    "chai": "^4.2.0",
-    "codecov": "^3.8.1",
-    "execa": "^5.0.0",
-    "gts": "^3.0.0",
-    "mocha": "^9.0.0",
-    "nock": "^13.0.0",
-    "pkg": "^5.0.0",
+    "@types/node": "^16.11.7",
+    "@types/server-destroy": "^1.0.1",
+    "@types/sinon": "^10.0.6",
+    "@types/update-notifier": "^5.1.0",
+    "c8": "^7.10.0",
+    "chai": "^4.3.4",
+    "codecov": "^3.8.2",
+    "execa": "^5.1.1",
+    "gts": "^3.1.0",
+    "mocha": "^9.1.3",
+    "nock": "^13.2.1",
+    "pkg": "^5.4.1",
     "semantic-release": "^18.0.0",
-    "sinon": "^12.0.0",
-    "strip-ansi": "^6.0.0",
-    "typescript": "^4.0.0"
+    "sinon": "^12.0.1",
+    "strip-ansi": "^7.0.1",
+    "typescript": "^4.4.4"
   },
   "engines": {
-    "node": ">=10"
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
   },
   "files": [
     "build/src"
